+++ CHANGELOG.md
@@ -9,7 +9,8 @@ This update focuses on the core of the Alchemy Engine. This allows for project l
 - Created the Shadowpaw.Alchemy namespace
 - Added Injector utility for dependency injection
 - Added Injector Contexts for scoped dependency injection
-- Added Configs type for project-level configurations
+- Added Service types for project-level services
+- Added Config types for project-level configs
 
 ## 1.0.0 - Initial Release (2024-07-25)